pyproject.toml
[tool.poetry]
name = "codeshare-back"
version = "0.1.0"
description = ""
authors = ["Neeraj319 <itskop520@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
python-dotenv = "^0.20.0"
aerich = "^0.6.3"
fastapi-pagination = "^0.9.3"
passlib = "^1.7.4"
python-jose = "^3.3.0"
uvicorn = "^0.17.6"
asyncpg = "^0.25.0"
bcrypt = "^3.2.2"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}


[tool.aerich]
tortoise_orm = "main.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
